[Subgraph] Update `SubGraphBase` to be virtual and Implement `SubGraphCpu`
- Updated the base class `SubGraphBase` to become a virtual base class and implemented a derived class `SubGraphCpu`.
- Set up `NetworkGraph` to default to creating and utilizing one instance of `SubGraphCpu`.
- Declared several member functions related to memory management and computation engines as pure virtual functions (=0) in `SubGraphBase`.
